+++ src/components/Menu/__tests__/Menu.spec.js
@@ -23,11 +23,35 @@ const getComponent = (props = {}) => mount(
 describe('<Menu />', () => {
   it('is closed by default and opens when the button was clicked', () => {
     const mounted = getComponent();
-    const menu = mounted.find('div').first();
-    expect(menu.hasClass('slds-is-open')).toBeFalsy();
-    expect(menu.find('Button').exists()).toBeTruthy();
-    menu.find('Button').simulate('click');
+    expect(mounted.find('.slds-dropdown-trigger').hasClass('slds-is-open')).toBeFalsy();
+    expect(mounted.find('Button').exists()).toBeTruthy();
+    mounted.find('Button').simulate('click');
     expect(mounted.find('.slds-dropdown-trigger').hasClass('slds-is-open')).toBeTruthy();
+  });
+
+  it('isOpen state override works', () => {
+    const mounted = getComponent({ isOpen: true });
+    expect(mounted.find('div').first().hasClass('is-open'));
+  });
+
+  it('uses size class', () => {
+    const mounted = getComponent({ size: 'small' });
+    expect(mounted.find('div').last().hasClass('slds-dropdown_small')).toBeTruthy();
+  });
+
+  it('uses left class', () => {
+    const mounted = getComponent({ position: 'top-left' });
+    expect(mounted.find('div').last().hasClass('slds-dropdown_left')).toBeTruthy();
+  });
+
+  it('uses right class', () => {
+    const mounted = getComponent({ position: 'top-right' });
+    expect(mounted.find('div').last().hasClass('slds-dropdown_right')).toBeTruthy();
+  });
+
+  it('uses bottom class', () => {
+    const mounted = getComponent({ position: 'bottom-right' });
+    expect(mounted.find('div').last().hasClass('slds-dropdown_bottom')).toBeTruthy();
   });
 
   it('starts as closed by default and defaultOpens', () => {